fix(classes): reverte idLote para long e infCad para List<infCad>
idLote: o schema TIdLote e [0-9]{1,15} (ate 15 digitos); int estoura
em ~2,1 bi (10 digitos). Restaura o long do fix #17442 (geracao de
idLote sem controle de concorrencia, baseada em timestamp).

infCad: o schema consCad define infCad com maxOccurs="unbounded";
restaura List<infCad> para capturar multiplos estabelecimentos no
retorno (consulta por Inscricao Unica).

+++ NFe.Classes/Servicos/Recepcao/enviNFe2.cs
@@ -38,7 +38,7 @@ namespace NFe.Classes.Servicos.Recepcao
     [XmlRoot(ElementName = "enviNFe", Namespace = "http://www.portalfiscal.inf.br/nfe")]
     public class enviNFe2
     {
-        public enviNFe2(string versao, int idLote, List<NFe> nFe)
+        public enviNFe2(string versao, long idLote, List<NFe> nFe)
         {
             this.versao = versao;
             this.idLote = idLote;
